Fix race condition on extra characteristic discovery in CoreBluetooth
Some devices are sending extra characteristic discovery events after
already being connected to the device. One such device is the GoCube
smart cube. If you attempt to subscribe to the data characteristic,
often an extra characteristic discovery event fires before the
subscription completes and wipes out the existing characteristic
data. When the subscription request completes, an attempt is made
to complete the subscription request in the subscribe_future_state
field of the characteristic. Unfortunately the extra discovery
event has cleared this field, so a panic results.